a folder for test has been added, and a test file for the Sensor has been Create
