+++ log.html
@@ -2081,31 +2081,31 @@ storage = function () {
 window.output = {};
 </script>
 <script type="text/javascript">
-window.output["suite"] = [1,2,3,0,[],[1,0,17],[],[[4,0,0,[],[1,15,1],[[0,5,6,0,0,0,0,0,[1,16,0],[]]]]],[],[1,1,0,0]];
+window.output["suite"] = [1,2,3,0,[],[1,0,32],[],[[4,0,0,[],[1,15,16],[[0,5,0,0,0,6,0,0,[1,16,15],[[0,7,8,0,0,9,0,0,[1,16,15],[[30,2,10]]]]]]]],[],[1,1,0,0]];
 </script>
 <script type="text/javascript">
 window.output["strings"] = [];
 </script>
 <script type="text/javascript">
-window.output["strings"] = window.output["strings"].concat(["*","*Testomar","*/home/omar/Desktop/Desktop-omar/Sesame2-sim/testomar.robot","*testomar.robot","*Start Test","*Start Sensor","*SensorSimulatorLibrary"]);
+window.output["strings"] = window.output["strings"].concat(["*","*SensorDataTest","*/home/omar/Desktop/Desktop-omar/Sesame2-sim/tests/SensorDataTest.robot","*tests/SensorDataTest.robot","*Test moving between the zones","*moving between the zones","*${live_data_moving_between_the_zones}","*Send Data","*SensorSimulatorLibrary","*${data}","*Received from server: {\"response\": \"OK\"}"]);
 </script>
 <script type="text/javascript">
-window.output["stats"] = [[{"elapsed":"00:00:00","fail":0,"label":"All Tests","pass":1,"skip":0}],[],[{"elapsed":"00:00:00","fail":0,"id":"s1","label":"Testomar","name":"Testomar","pass":1,"skip":0}]];
+window.output["stats"] = [[{"elapsed":"00:00:00","fail":0,"label":"All Tests","pass":1,"skip":0}],[],[{"elapsed":"00:00:00","fail":0,"id":"s1","label":"SensorDataTest","name":"SensorDataTest","pass":1,"skip":0}]];
 </script>
 <script type="text/javascript">
 window.output["errors"] = [];
 </script>
 <script type="text/javascript">
-window.output["baseMillis"] = 1715952330923;
+window.output["baseMillis"] = 1716201676161;
 </script>
 <script type="text/javascript">
-window.output["generated"] = 18;
+window.output["generated"] = 34;
 </script>
 <script type="text/javascript">
 window.output["expand_keywords"] = null;
 </script>
 <script type="text/javascript">
-window.settings = {"defaultLevel":"INFO","minLevel":"NONE","reportURL":"report.html","rpa":false,"splitLogBase":"log","title":""};
+window.settings = {"defaultLevel":"INFO","minLevel":"INFO","reportURL":"report.html","rpa":false,"splitLogBase":"log","title":""};
 </script>
 <title></title>
 </head>
